+++ src/types/composed/index.ts
@@ -1,5 +1,0 @@
-export type { ErrorType } from '../errors';
-export type { Brand } from './brand';
-export type { Identity } from './identity';
-export type { Trait } from './trait';
-export type { Transformation } from './transformation';