+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## v1.0.5 2026-02-08
 
-### Changed
+### Added
 
-- Compiled code with tsup
+- Single assertor import
 
 ## v1.0.4 2026-02-07